Reverse Dithering
These are a few changes to the original rosetta code for dithering for two types of upgrades to the dithering process BEFORE the actual dithering is run. First is a server side, ie prequantisation step. And the second is an optional client side step. It is fairly unique in it's approach as it assumes that even the original was the result of some sort of dithering and adds colors before conducting the standard dithering. This gives good quality dithering.

Additional changes in the dithering option unrelated to reverse dithering, have to do with saving the color template and using a color template etc.